backend/pyproject.toml
[project]
name = "langgraph-orchestrator"
version = "0.1.0"
description = "Session orchestrator for MecanoLabs/MecanoConsulting work prioritization"
authors = [
    { name = "Samir Awuapara", email = "samir@mecanolabs.com" }
]
requires-python = ">=3.10"
dependencies = [
    # LangGraph + LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",

    # API Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",

    # Database for checkpoints
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
orchestrator = "src.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
